Fixed NPE for disabled cache

+++ src/main/java/ch/o2it/weblounge/cache/impl/CacheConfigurationFactory.java
@@ -256,7 +256,8 @@ public class CacheConfigurationFactory implements ManagedService {
       return;
     
     Configuration config = configHolder.getConfiguration();
-    config.delete();
+    if (config != null)
+      config.delete();
   }
 
   /**
